requirements.txt
cloudpickle==3.1.0
deep-forest @ git+https://github.com/LAMDA-NJU/Deep-Forest.git
lightgbm==4.6.0
joblib==1.4.2
keras==2.15.0
numba==0.61.0
numpy==1.26.4
openpyxl==3.1.5
optuna==4.2.1
pandas==2.2.3
psycopg2-binary==2.9.10
pyarrow==19.0.0
rpy2==3.5.17
scikit-learn==1.5.2
scipy==1.16.3
seaborn==0.13.2
smote-variants==1.0.1
tabulate==0.9.0
tensorflow-cpu==2.15.0
threadpoolctl==3.5.0
